+++ lib/src/shared_map_extension.dart
@@ -31,6 +31,22 @@ extension FutureSharedMapExtension<K, V> on Future<SharedMap<K, V>> {
 
   Future<SharedMapCached<K, V>> cached({Duration? timeout}) =>
       then((o) => o.cached(timeout: timeout));
+
+  Future<SharedMap<K, V>> setCallbacks(
+          {SharedMapEntryCallback<K, V>? onPut,
+          SharedMapEntryCallback<K, V>? onRemove}) =>
+      then((o) {
+        o.setCallbacks(onPut: onPut, onRemove: onRemove);
+        return o;
+      });
+
+  Future<SharedMap<K, V>> setCallbacksDynamic<K1, V1>(
+          {SharedMapEntryCallback<K1, V1>? onPut,
+          SharedMapEntryCallback<K1, V1>? onRemove}) =>
+      then((o) {
+        o.setCallbacksDynamic(onPut: onPut, onRemove: onRemove);
+        return o;
+      });
 }
 
 /// Extension on [FutureOr]<[SharedMap]`<K, V>`>
@@ -140,6 +156,36 @@ extension FutureOrSharedMapExtension<K, V> on FutureOr<SharedMap<K, V>> {
       return self.cached(timeout: timeout);
     } else {
       return self.cached(timeout: timeout);
+    }
+  }
+
+  FutureOr<SharedMap<K, V>> setCallbacks(
+      {SharedMapEntryCallback<K, V>? onPut,
+      SharedMapEntryCallback<K, V>? onRemove}) {
+    var self = this;
+    if (self is Future<SharedMap<K, V>>) {
+      return self.then((o) {
+        o.setCallbacks(onPut: onPut, onRemove: onRemove);
+        return o;
+      });
+    } else {
+      self.setCallbacks(onPut: onPut, onRemove: onRemove);
+      return self;
+    }
+  }
+
+  FutureOr<SharedMap<K, V>> setCallbacksDynamic<K1, V1>(
+      {SharedMapEntryCallback<K1, V1>? onPut,
+      SharedMapEntryCallback<K1, V1>? onRemove}) {
+    var self = this;
+    if (self is Future<SharedMap<K, V>>) {
+      return self.then((o) {
+        o.setCallbacksDynamic(onPut: onPut, onRemove: onRemove);
+        return o;
+      });
+    } else {
+      self.setCallbacksDynamic(onPut: onPut, onRemove: onRemove);
+      return self;
     }
   }
 }